requirements.txt
bcrypt
Django
djangorestframework
enum34
jsonschema
passlib
PyJWT